crud for expenses created, removed userId query parameter, and about to implement clientside keys

+++ app/src/main/java/com/example/wallettracker/data/expense/ExpenseResponse.kt
@@ -5,5 +5,5 @@ data class ExpenseResponse(
     val expenseDate: String,
     val id: Long,
     val price: Double,
-    val user: Long
+    val user: Int
 )
